Created/copied a dev shell implementation for bootstrapping
Thanks to Misterio77's rep: https://github.com/Misterio77/nix-config/tree/main

Don't quite understand the syntax and if for each system stuff has to be
there, but works for now. Haven't tested on a fresh system without git,
flake support, etc, to see if a valid way to "bootstrap". Need to figure
out validating that.